+++ packages/sit-onyx/src/components/OnyxForm/OnyxForm.stories.ts
@@ -2,6 +2,7 @@ import { withNativeEventLogging } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
 import { h } from "vue";
 import { OnyxInput, OnyxToast } from "../..";
+import { ShowErrorModes } from "../../composables/useErrorClass";
 import FormExample from "../examples/FormExample/FormExample.vue";
 import FormExampleSourceCode from "../examples/FormExample/FormExample.vue?raw";
 import OnyxButton from "../OnyxButton/OnyxButton.vue";
@@ -15,6 +16,12 @@ import OnyxForm from "./OnyxForm.vue";
 const meta: Meta<typeof OnyxForm> = {
   title: "Form Elements/Form",
   component: OnyxForm,
+  argTypes: {
+    showError: {
+      control: "select",
+      options: ShowErrorModes,
+    },
+  },
 };
 
 export default meta;
@@ -27,7 +34,7 @@ export const Default = {
   args: {
     style: { maxWidth: "10rem", display: "flex", flexDirection: "column", gap: "1rem" },
     default: () => [
-      h(OnyxInput, { label: "Favorite band", modelValue: "Queen" }),
+      h(OnyxInput, { label: "Favorite band", modelValue: "Que2en", pattern: "[A-Za-z ]+" }),
       h(OnyxInput, { label: "Favorite password", type: "password", modelValue: "incorrect" }),
       h(OnyxStepper, { label: "Number of hairs", min: 0, modelValue: 23 }),
       h(
@@ -47,6 +54,34 @@ export const Default = {
   },
 } satisfies Story;
 
+/**
+ * This example show a form that displays all errors immediately.
+ */
+export const ShowError = {
+  args: {
+    ...Default.args,
+    default: () => [
+      h(OnyxInput, { label: "Favorite band", modelValue: "2. Queen", pattern: "[A-Za-z ]+" }),
+      h(OnyxInput, { label: "Favorite password", type: "password", modelValue: "incorrect" }),
+      h(OnyxStepper, { label: "Number of hairs", min: 0, modelValue: 23 }),
+      h(
+        "div",
+        {
+          style: { display: "flex", gap: "0.5rem" },
+        },
+        [
+          h(OnyxButton, { label: "Reset", type: "reset", mode: "outline" }),
+          h(OnyxButton, { label: "Submit", type: "submit", formmethod: "dialog" }), // we use formmethod `dialog` to avoid a page load on submit
+        ],
+      ),
+    ],
+    showError: true,
+  },
+} satisfies Story;
+
+/**
+ * This example show a form that is disabled as a whole.
+ */
 export const Disabled = {
   args: {
     ...Default.args,