+++ packages/sit-onyx/src/components/OnyxInput/types.ts
@@ -1,11 +1,13 @@
 import type { DensityProp } from "../../composables/density";
 import type { RequiredMarkerProp } from "../../composables/required";
 import type { CustomValidityProp } from "../../composables/useCustomValidity";
+import type { SkeletonInjected } from "../../composables/useSkeletonState";
 import type { AutofocusProp } from "../../types";
-import type { FormInjected } from "../OnyxForm/OnyxForm.core";
+import type { FormInjectedProps } from "../OnyxForm/OnyxForm.core";
 import type { OnyxFormElementProps } from "../OnyxFormElement/types";
 
-export type OnyxInputProps = DensityProp &
+export type OnyxInputProps = FormInjectedProps &
+  DensityProp &
   RequiredMarkerProp &
   CustomValidityProp &
   Omit<OnyxFormElementProps, "modelValue" | "errorMessages"> &
@@ -53,10 +55,6 @@ export type OnyxInputProps = DensityProp &
      */
     readonly?: boolean;
     /**
-     * Whether the input should be disabled.
-     */
-    disabled?: FormInjected<boolean>;
-    /**
      * Whether the input is loading. User interaction will be disabled.
      */
     loading?: boolean;
@@ -70,7 +68,7 @@ export type OnyxInputProps = DensityProp &
     /**
      * Whether to show a skeleton input.
      */
-    skeleton?: boolean;
+    skeleton?: SkeletonInjected;
   };
 
 export const INPUT_TYPES = ["email", "password", "search", "tel", "text", "url"] as const;