+++ packages/sit-onyx/src/components/OnyxStepper/types.ts
@@ -1,11 +1,13 @@
 import type { DensityProp } from "../../composables/density";
 import type { CustomValidityProp } from "../../composables/useCustomValidity";
+import type { SkeletonInjected } from "../../composables/useSkeletonState";
 import type { AutofocusProp } from "../../types";
-import type { FormInjected } from "../OnyxForm/OnyxForm.core";
+import type { FormInjectedProps } from "../OnyxForm/OnyxForm.core";
 import type { OnyxFormElementProps } from "../OnyxFormElement/types";
 import type { Autocomplete } from "../OnyxInput/types";
 
-export type OnyxStepperProps = DensityProp &
+export type OnyxStepperProps = FormInjectedProps &
+  DensityProp &
   CustomValidityProp &
   Omit<OnyxFormElementProps, "modelValue" | "errorMessages" | "withCounter" | "maxlength"> &
   AutofocusProp & {
@@ -52,10 +54,6 @@ export type OnyxStepperProps = DensityProp &
      */
     autocomplete?: Autocomplete;
     /**
-     * Whether to disable the input and prevent user interaction.
-     */
-    disabled?: FormInjected<boolean>;
-    /**
      * Whether the input should be readonly.
      */
     readonly?: boolean;
@@ -66,5 +64,5 @@ export type OnyxStepperProps = DensityProp &
     /**
      * Whether to show a skeleton input.
      */
-    skeleton?: boolean;
+    skeleton?: SkeletonInjected;
   };