+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-beta.55
+
+### Patch Changes
+
+- Updated dependencies [772f6a5]
+- Updated dependencies [772f6a5]
+  - sit-onyx@1.0.0-beta.55
+
+## 1.0.0-beta.54
+
+### Patch Changes
+
+- Updated dependencies [01f8664]
+  - sit-onyx@1.0.0-beta.54
+
+## 1.0.0-beta.53
+
+### Patch Changes
+
+- sit-onyx@1.0.0-beta.53
+
 ## 1.0.0-beta.52
 
 ### Patch Changes